requirements.txt
# YouTube API
google-api-python-client>=2.100.0
google-auth>=2.23.0

# CLI & UI
click>=8.1.0
rich>=13.0.0

# Environment & Utils
python-dotenv>=1.0.0
pytz>=2024.1

# Database
sqlite-utils>=3.35.0
